Contains dictionary of possible foods.
Explanation of the format of the dictionary is in the file.  Some of the entries were split up / rewritten to be more 'speech' friendly.  For example, instead of 'lemon/citrus' as an entry, the dictionary has 'lemon', 'citrus', and 'lemon or citrus' each as an entry with the same 'id' value.  Another example is 'pasta & grains', which is represented as 'pasta and grains' in the dictionary.